+++ org.eclipse.ajdt.ui.tests/src/org/eclipse/ajdt/ui/tests/javamodel/Bug117327Test.java
@@ -1,10 +1,10 @@
 /*******************************************************************************
  * Copyright (c) 2004 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials 
+ * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- * 
+ *
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *     Sian January - initial version
@@ -32,15 +32,15 @@ public class Bug117327Test extends UITestCase {
 					IResource.DEPTH_INFINITE);
 		assertEquals("Expected two declaration markers", 2, markers.length); //$NON-NLS-1$
 		Integer lineNumber = (Integer)markers[0].getAttribute(IMarker.LINE_NUMBER);
-		assertTrue("Expected a declaration marker at line 15 or 18", 15 == lineNumber.intValue() || 18 == lineNumber.intValue()); //$NON-NLS-1$
-		if(lineNumber.intValue() == 15) {
+		assertTrue("Expected a declaration marker at line 15 or 18", 15 == lineNumber || 18 == lineNumber); //$NON-NLS-1$
+		if(lineNumber == 15) {
 			lineNumber = (Integer)markers[1].getAttribute(IMarker.LINE_NUMBER);
-			assertEquals("Expected a declaration marker at line 18", 18, lineNumber.intValue()); //$NON-NLS-1$				
+			assertEquals("Expected a declaration marker at line 18", 18, lineNumber.intValue()); //$NON-NLS-1$
 		} else {
 			lineNumber = (Integer)markers[1].getAttribute(IMarker.LINE_NUMBER);
-			assertEquals("Expected a declaration marker at line 15", 15, lineNumber.intValue()); //$NON-NLS-1$	
+			assertEquals("Expected a declaration marker at line 15", 15, lineNumber.intValue()); //$NON-NLS-1$
 		}
 	}
-	
-	
+
+
 }
